Initial port of Searchable from Ministudy.
This is the first 'usable' release, incorporating the skeleton of the SQL
query builder from Ministudy. Additional functionality for performing
queries and using the aggregation framework for MongoDB has been added, as
well as a reworking of the grammar to handle edge cases.

This commit is pre-release, and is currently only suitable for MongoDB.